Extracted StateWrapper into its own file.
